requirements.txt
absl-py==1.4.0
adblockparser==0.7
astor==0.8.1
astunparse==1.6.3
attrs==23.1.0
Automat==22.10.0
billiard==3.6.4.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
crochet==2.0.0
cryptography==41.0.1
cssselect==1.2.0
docker==6.1.3
execnet==1.9.0
filelock==3.12.2
flatbuffers==23.5.26
funcparserlib==1.0.1
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h2==4.1.0
h5py==3.9.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.1.0
jax==0.4.12
jmespath==1.0.1
keras==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
parsel==1.8.1
Pillow==9.5.0
playwright==1.35.0
pluggy==1.2.0
praw==7.7.0
prawcore==2.3.0
priority==1.3.0
Protego==0.2.1
protobuf==4.23.3
psutil==5.9.5
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==9.0.4
pyOpenSSL==23.2.0
pytest==7.3.2
pytest-forked==1.6.0
pytest-xdist==3.3.1
python-dotenv==1.0.0
qt5reactor==0.6.3
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
Scrapy==2.9.0
scrapy-playwright==0.0.26
scrapy-splash==0.9.0
scrapyscript==1.1.5
service-identity==21.1.0
six==1.16.0
splash==3.5
tensorboard==1.14.0
tensorboard-data-server==0.7.1
tensorflow-estimator==1.14.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tldextract==3.4.4
Twisted==22.10.0
typing_extensions==4.6.3
update-checker==0.18.0
urllib3==1.26.16
w3lib==2.1.1
websocket-client==1.5.3
Werkzeug==2.3.6
wrapt==1.14.1
xvfbwrapper==0.2.9
zope.interface==6.0
